cmd/snap-confine: fold snap-confine-invocation.[ch] into snap-confine.[ch]
